# Slim down and cache variant resolving artifact set instances