--- a/platforms/software/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/resolveengine/artifact/ArtifactSelectionServices.java
+++ b/platforms/software/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/resolveengine/artifact/ArtifactSelectionServices.java
@@ -20,6 +20,9 @@ import org.gradle.api.internal.artifacts.transform.ArtifactVariantSelector;
 import org.gradle.api.internal.artifacts.transform.ResolvedVariantTransformer;
 import org.gradle.api.internal.artifacts.transform.TransformUpstreamDependenciesResolver;
 import org.gradle.api.internal.artifacts.transform.TransformedVariantFactory;
+import org.gradle.api.internal.attributes.immutable.ImmutableAttributesSchema;
+import org.gradle.internal.component.model.GraphVariantSelector;
+import org.gradle.internal.resolve.resolver.VariantArtifactResolver;
 
 /**
  * Services provided by the consumer to be used during artifact selection.
@@ -28,14 +31,23 @@ public class ArtifactSelectionServices {
 
     private final ArtifactVariantSelector variantSelector;
     private final ResolvedVariantTransformer resolvedVariantTransformer;
+    private final VariantArtifactResolver variantResolver;
+    private final GraphVariantSelector graphVariantSelector;
+    private final ImmutableAttributesSchema consumerSchema;
 
     public ArtifactSelectionServices(
         ArtifactVariantSelector variantSelector,
         TransformedVariantFactory transformedVariantFactory,
-        TransformUpstreamDependenciesResolver dependenciesResolver
+        TransformUpstreamDependenciesResolver dependenciesResolver,
+        VariantArtifactResolver variantResolver,
+        GraphVariantSelector graphVariantSelector,
+        ImmutableAttributesSchema consumerSchema
     ) {
         this.variantSelector = variantSelector;
         this.resolvedVariantTransformer = new ResolvedVariantTransformer(transformedVariantFactory, dependenciesResolver);
+        this.variantResolver = variantResolver;
+        this.graphVariantSelector = graphVariantSelector;
+        this.consumerSchema = consumerSchema;
     }
 
     public ArtifactVariantSelector getArtifactVariantSelector() {
@@ -45,4 +57,17 @@ public class ArtifactSelectionServices {
     public ResolvedVariantTransformer getResolvedVariantTransformer() {
         return resolvedVariantTransformer;
     }
+
+    public VariantArtifactResolver getVariantArtifactResolver() {
+        return variantResolver;
+    }
+
+    public GraphVariantSelector getGraphVariantSelector() {
+        return graphVariantSelector;
+    }
+
+    public ImmutableAttributesSchema getConsumerSchema() {
+        return consumerSchema;
+    }
+
 }