--- a/platforms/software/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/resolveengine/artifact/ResolvedArtifactsGraphVisitor.java
+++ b/platforms/software/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/resolveengine/artifact/ResolvedArtifactsGraphVisitor.java
@@ -16,53 +16,46 @@
 
 package org.gradle.api.internal.artifacts.ivyservice.resolveengine.artifact;
 
-import it.unimi.dsi.fastutil.longs.Long2ObjectMap;
-import it.unimi.dsi.fastutil.longs.Long2ObjectOpenHashMap;
+import org.gradle.api.artifacts.capability.CapabilitySelector;
+import org.gradle.api.internal.artifacts.ivyservice.resolveengine.excludes.specs.ExcludeSpec;
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.graph.DependencyGraphEdge;
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.graph.DependencyGraphNode;
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.graph.DependencyGraphVisitor;
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.graph.RootGraphNode;
 import org.gradle.api.internal.artifacts.type.ArtifactTypeRegistry;
-import org.gradle.api.internal.attributes.immutable.ImmutableAttributesSchema;
+import org.gradle.api.internal.attributes.ImmutableAttributes;
 import org.gradle.internal.component.local.model.LocalFileDependencyMetadata;
 import org.gradle.internal.component.model.ComponentGraphResolveState;
-import org.gradle.internal.component.model.GraphVariantSelector;
+import org.gradle.internal.component.model.IvyArtifactName;
 import org.gradle.internal.component.model.VariantGraphResolveState;
 import org.gradle.internal.model.CalculatedValueContainerFactory;
-import org.gradle.internal.resolve.resolver.ArtifactResolver;
-import org.gradle.internal.resolve.resolver.DefaultVariantArtifactResolver;
-import org.gradle.internal.resolve.resolver.VariantArtifactResolver;
 
-import java.util.function.LongFunction;
+import java.util.List;
+import java.util.Set;
 
 /**
  * Adapts a {@link DependencyArtifactsVisitor} to a {@link DependencyGraphVisitor}. Calculates the artifacts contributed by each edge in the graph and forwards the results to the artifact visitor.
  */
 public class ResolvedArtifactsGraphVisitor implements DependencyGraphVisitor {
-    private int nextId;
-    private final Long2ObjectMap<ArtifactsForNode> artifactsByNodeId = new Long2ObjectOpenHashMap<>();
+
     private final DependencyArtifactsVisitor artifactResults;
     private final ArtifactTypeRegistry artifactTypeRegistry;
+    private final VariantArtifactSetCache artifactSetCache;
     private final CalculatedValueContainerFactory calculatedValueContainerFactory;
-    private final VariantArtifactResolver variantResolver;
-    private final GraphVariantSelector graphVariantSelector;
-    private final ImmutableAttributesSchema consumerSchema;
+
+    // State
+    private int nextId;
 
     public ResolvedArtifactsGraphVisitor(
         DependencyArtifactsVisitor artifactsBuilder,
         ArtifactTypeRegistry artifactTypeRegistry,
-        CalculatedValueContainerFactory calculatedValueContainerFactory,
-        ArtifactResolver artifactResolver,
-        ResolvedVariantCache resolvedVariantCache,
-        GraphVariantSelector graphVariantSelector,
-        ImmutableAttributesSchema consumerSchema
+        VariantArtifactSetCache artifactSetCache,
+        CalculatedValueContainerFactory calculatedValueContainerFactory
     ) {
         this.artifactResults = artifactsBuilder;
         this.artifactTypeRegistry = artifactTypeRegistry;
+        this.artifactSetCache = artifactSetCache;
         this.calculatedValueContainerFactory = calculatedValueContainerFactory;
-        this.variantResolver = new DefaultVariantArtifactResolver(artifactResolver, artifactTypeRegistry, resolvedVariantCache);
-        this.graphVariantSelector = graphVariantSelector;
-        this.consumerSchema = consumerSchema;
     }
 
     @Override
@@ -72,15 +65,7 @@ public class ResolvedArtifactsGraphVisitor implements DependencyGraphVisitor {
 
     @Override
     public void visitEdges(DependencyGraphNode node) {
-        boolean hasTransitiveIncomingEdge = false;
-        for (DependencyGraphEdge dependency : node.getIncomingEdges()) {
-            hasTransitiveIncomingEdge |= dependency.isTransitive();
-            if (dependency.contributesArtifacts()) {
-                DependencyGraphNode parent = dependency.getFrom();
-                ArtifactsForNode artifacts = resolveVariantArtifacts(dependency, node);
-                artifactResults.visitArtifacts(parent, node, artifacts.artifactSetId, artifacts.artifactSet);
-            }
-        }
+        boolean hasTransitiveIncomingEdge = visitNonFileEdges(node);
 
         if (node.isRoot() || hasTransitiveIncomingEdge) {
             // Since file dependencies are not modeled as actual edges, we need to verify
@@ -92,38 +77,76 @@ public class ResolvedArtifactsGraphVisitor implements DependencyGraphVisitor {
         }
     }
 
+    /**
+     * Visit all the non-file edges for the given node.
+     *
+     * @return true if there is a transitive incoming edge.
+     */
+    private boolean visitNonFileEdges(DependencyGraphNode node) {
+        boolean hasTransitiveIncomingEdge = false;
+
+        int implicitArtifactSetId = -1;
+        ArtifactSet implicitArtifactSet = null;
+
+        for (DependencyGraphEdge edge : node.getIncomingEdges()) {
+            hasTransitiveIncomingEdge |= edge.isTransitive();
+
+            if (edge.contributesArtifacts()) {
+                if (maybeVisitAdhocEdge(node, edge)) {
+                    // The artifacts for this node were modified by the dependency.
+                    // Do not use the implicit artifact set.
+                    continue;
+                }
+
+                // Since the dependency does not modify the artifacts, we can use the same
+                // artifact set as other dependencies that do not modify the artifacts. We call
+                // this the implicit artifact set.
+                if (implicitArtifactSet == null) {
+                    // We have not visited the implicit artifacts yet.
+                    implicitArtifactSetId = nextId++;
+                    implicitArtifactSet = artifactSetCache.getImplicitVariant(node.getOwner().getResolveState(), node.getResolveState());
+                }
+
+                artifactResults.visitArtifacts(edge.getFrom(), node, implicitArtifactSetId, implicitArtifactSet);
+            }
+        }
+
+        return hasTransitiveIncomingEdge;
+    }
+
+    /**
+     * Process the given edge, and if it modifies the artifacts, visit the artifacts.
+     *
+     * @return true if this edge modifies the artifacts, meaning it is adhoc, and should
+     * not also contribute to the implicit artifact set.
+     */
+    private boolean maybeVisitAdhocEdge(DependencyGraphNode node, DependencyGraphEdge dependency) {
+        ComponentGraphResolveState component = node.getOwner().getResolveState();
+        VariantGraphResolveState variant = node.getResolveState();
+
+        ImmutableAttributes attributes = dependency.getAttributes();
+        List<IvyArtifactName> artifacts = dependency.getDependencyMetadata().getArtifacts();
+        ExcludeSpec exclusions = dependency.getExclusions();
+        Set<CapabilitySelector> capabilitySelectors = dependency.getDependencyMetadata().getSelector().getCapabilitySelectors();
+
+        // If all dependency modifiers are empty, this edge does not produce an adhoc artifact set.
+        if (artifacts.isEmpty() &&
+            attributes.isEmpty() &&
+            capabilitySelectors.isEmpty() &&
+            !exclusions.mayExcludeArtifacts()
+        ) {
+            return false;
+        }
+
+        int id = nextId++;
+        VariantResolvingArtifactSet artifactSet = new VariantResolvingArtifactSet(component, variant, attributes, artifacts, exclusions, capabilitySelectors);
+        artifactResults.visitArtifacts(dependency.getFrom(), node, id, artifactSet);
+
+        return true;
+    }
+
     @Override
     public void finish(RootGraphNode root) {
         artifactResults.finishArtifacts(root);
-        artifactsByNodeId.clear();
-    }
-
-    private ArtifactsForNode resolveVariantArtifacts(DependencyGraphEdge dependency, DependencyGraphNode toNode) {
-        ComponentGraphResolveState component = toNode.getOwner().getResolveState();
-        VariantGraphResolveState variant = toNode.getResolveState();
-
-        // Do not share an ArtifactSet if the artifacts are modified by the dependency.
-        if (!dependency.getDependencyMetadata().getArtifacts().isEmpty() ||
-            !dependency.getAttributes().isEmpty() ||
-            dependency.getExclusions().mayExcludeArtifacts()
-        ) {
-            int id = nextId++;
-            return new ArtifactsForNode(id, new VariantResolvingArtifactSet(variantResolver, component, variant, dependency, graphVariantSelector, consumerSchema, calculatedValueContainerFactory));
-        }
-
-        return artifactsByNodeId.computeIfAbsent(toNode.getNodeId(), (LongFunction<ArtifactsForNode>) value -> {
-            int id = nextId++;
-            return new ArtifactsForNode(id, new VariantResolvingArtifactSet(variantResolver, component, variant, dependency, graphVariantSelector, consumerSchema, calculatedValueContainerFactory));
-        });
-    }
-
-    private static class ArtifactsForNode {
-        private final int artifactSetId;
-        private final ArtifactSet artifactSet;
-
-        ArtifactsForNode(int artifactSetId, ArtifactSet artifactSet) {
-            this.artifactSetId = artifactSetId;
-            this.artifactSet = artifactSet;
-        }
     }
 }