… platforms/software/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/resolveengine/artifact/VariantArtifactSetCache.java
… platforms/software/dependency-management/src/main/java/org/gradle/api/internal/artifacts/repositories/metadata/DefaultMavenAttributesFactory.java
… platforms/software/dependency-management/src/main/java/org/gradle/api/internal/artifacts/transform/ConsumerProvidedVariantFinder.java
… platforms/software/dependency-management/src/main/java/org/gradle/api/internal/attributes/matching/DefaultAttributeMatcher.java
… platforms/software/dependency-management/src/test/groovy/org/gradle/api/internal/artifacts/ivyservice/resolveengine/artifact/DefaultVisitedArtifactResultsTest.groovy
… platforms/software/dependency-management/src/test/groovy/org/gradle/api/internal/artifacts/ivyservice/resolveengine/artifact/VariantResolvingArtifactSetTest.groovy
… platforms/software/dependency-management/src/test/groovy/org/gradle/api/internal/artifacts/transform/ConsumerProvidedVariantFinderTest.groovy
… subprojects/core/src/main/java/org/gradle/api/internal/attributes/AbstractAttributesFactory.java
… subprojects/core/src/main/java/org/gradle/api/internal/attributes/DefaultAttributesFactory.java
… subprojects/core/src/main/java/org/gradle/api/internal/attributes/DefaultMutableAttributeContainer.java