This PR contains a number of optimizations discovered while profiling and analyzing heap dumps. 

Most importantly, it includes a commit which greatly reduces the number of instances of `VariantResolvingArtifactSet` we create, and further decreases the memory footprint of the instances we do create. 

Altogether, these changes result in a 16% execution time improvement and a 12%-19% heap usage improvement when syncing the android extra large 6000 project benchmark. The sync time was reduced from ~3.5 minutes to ~3 minutes, and the required heap reduced from ~26gb to ~22gb.

### Reviewing cheatsheet

Before merging the PR, comments starting with 
- ❌ ❓**must** be fixed
- 🤔 💅 **should** be fixed
- 💭 **may** be fixed
- 🎉 celebrate happy things