@jvandort 2024-09-24
After performance and memory testing, these changes show an 8 to 9 percent performance improvement in the Android extra large 4000+ project studio sync benchmark. The benchmark was updated to use modern versions of AGP and KGP

![image](https://github.com/user-attachments/assets/deda7b69-fa9a-4554-a254-2ad6120e1e1e)

Furthermore, with the changes the 4000+ project sync will succeed with as little as 14gb of memory. Before these changes, the project would not sync below 16gb.