@jvandort 2024-09-24
@bot-gradle test this