@jvandort 2024-09-24
How does this interact with component metadata rules?