@jvandort 2024-09-25
We may need to ease up on some of the caching here, but there are useful optimizations here that work