@jvandort 2024-11-01
This PR shows the following performance improvement (with the profiler enabled)

Before Slim:

Min: 211.5 s
Max: 226.1 s
Mean: 218.9 s
Median: 219.0 s
Standard Deviation: 5.6 s
After Slim:

Min: 181.5 s
Max: 188.1 s
Mean: 183.5 s
Median: 183.4 s
Standard Deviation: 1.9 s
Improvements:

Mean Improvement: 35.4 s (16.2%)
Median Improvement: 35.6 s (16.3%)