@jvandort 2024-11-01
I manually tested this PR for memory usage.

Prior to the changes, the 6000 project android extra large benchmark would sync with no less than 26gb of heap. 

Following the changes, the 6000 project android extra large bench mark would sync with no less than 22gb of heap.

Given the low precision of these heap values, this means the required heap value could have improved anywhere from 12% to 19%