@jvandort 2024-11-01
> How does this interact with component metadata rules?

Component metadata rules are applied to mutable component _metadata_ instances. This cache operates on component _state_ instances, which are derived from metadata instances. The component and variant state types created from different component metadata instances (potentially with different rules applied to each) will result in component and variant state instances with differing instance IDs. 

The caching implemented in this PR _would_ work for components modified my CMRs, however the cache is inefficient even for non-modified external components, since each repository creates its own component metadata and thus state object.

The cache is not enabled for external components due to this inefficiency, even though it would technically work, even with CMRs applied. Future work in this area should focus on deduplicating repositories so that they can share component metadata/state instances.